Update rubyzip to latest version, fixing `can't dup NilClass`
`rubyzip` fails on some files (specifically tested zip files created with Rust). This means the`CodeDeploy` agent always failson those zips. The zip is valid and Finder on macOS unzips it fine.

The bug appears to already be fixed in the latest `rubyzip`:

* https://github.com/rubyzip/rubyzip/issues/82

More than a couple people are seeing this issue:

* https://forums.aws.amazon.com/thread.jspa?threadID=210427&tstart=0
* https://discuss.circleci.com/t/aws-codedeploy-cant-dup-nilclass-in-downloadbundle/3014