[PATCH 12/49] Adding DeploymentGroupId as a required parameter to the CodeDeploy Local CLI
This is due to the fact that if people ever have a real CodeDeploy
deployment they would want to correctly deploy a revision locally that
uses the ApplicationStop from the previous revision. The only way to do
that is to ask for the deployment group id. If we don't do this then if
they have two applications on the same host they might stomp on each
other if we use the same deployment group id folder very time. But if we
use a different random one very time that also wouldn't satisfy all
customers. The best solution is to require it and be transparent about
what it entails.

cr https://code.amazon.com/reviews/CR-183052