Define `open_timeout` for `get_ec2_metadata_region`
While developing some AMI provisioning tools locally, we attempt to
install the CodeDeploy agent. Unfortunately, the process was hanging
indefinitely.

Digging into the process using `strace`, it is stuck on a `connect` to
the EC2 metadata service.

```
connect(7, {sa_family=AF_INET, sin_port=htons(80), sin_addr=inet_addr("169.254.169.254")}, 16
```

This is due to the metadata service not being present in our local
development environment and the installer isn't able to resolve the
service.

After locating this in the installation script, I can see there is a
(very high) read timeout however there isn't an open timeout which is
what the process will use to restrict the amount of time a connection
can be pending a connection. This is particularly useful if the service
you are attempting to connect to is either unresponsive to accept the
connection or not reachable due to things like DNS resolution.

To address the issue, I've added an open timeout to the `open` call and
now the process fails as expected after a sensible 3s.