Issue #185:
Makes codedeploy-agent compatible with Ubuntu 18.04 when deploying two different apps in the base common directory. ie.:
- `/var/www/some-api/`
- `/var/www/some-angular-app/`

This works fine on Ubuntu 16.04, but you get the following error on Ubuntu 18.04 with ruby 2.5:
`Unknown error, Directory not empty @ dir_s_rmdir - /var/www`

It might not be the best way to do it, but it is meant to provide a solution until this officially gets fixed since it has been open since August 2018 already.