[Ruby SDK v3 upgrade] Remove deprecated underscore method call
Prior to this change, the Agent was using aws sdk ruby v2. That version had a method called Seahorse::Util.underscore(), which was used in the Agent. We are in the process of upgrading the agent to use v3 of the SDK, in which that method does not exist.

This change replaces that Seahorse::Util.underscore() method call with a newly created method that does the same thing. I also added validation to make sure the commands being passed in to the underscore method call were in the PascalForm, as that is what the Agent commands are always in. We need this update so that when we upgrade to v3, the code will be compatible.

 * Unit Tests : [Y]
 * Integration Tests : [N]

Seahorse::Util.underscore method: https://www.rubydoc.info/gems/aws-sdk-core/2.9.44/Seahorse%2FUtil.underscore