[Ruby SDK v3 Upgrade] Replace deprecated add_service method and upgrade to v3
Prior to this change, the ruby sdk v3 did not include the Aws.add_service method, which we used inside the agent. The ruby sdk v3 also split up the aws service clients into separate gems.

This change replaces the Aws.add_service method with the implementation found in the AwsSdkRubyCodeGenWrapper
It also includes the aws-sdk-s3 gem to replace all of the dependencies on 'aws-sdk-core/s3' with 'aws-sdk-s3'.

 * Unit Tests : [Y]
 * Integration Tests : [N]