[Agent 1.1.2] Add unit test to verify unpack_bundle ruby unzip mechanism
Why is this change needed?
---------
Prior to this change, we didn't have a way to test the "overwrite" logic
in ruby unzip during the fallback mechanism of unpack_bundle method in
DownloadBundle command.

How does it address the issue?
---------
This change adds a unit test to verify if the fallback mechanism works
with a debris file in the deployment archive directory.