[ISO Release] Add endpoint support for ADCS, S3, and S3 in the install/update scripts.
Prior to this change, CodeDeploy endpoints did not have support for ISO
regions. S3 endpoints did not rely on the AWS SDK when it could have
for getting those endpoints. Install/Update scripts did not have ISO
region support.

This change uses the instance metadata to get the CodeDeploy endpoint.
Uses the SDK to retrieve the S3 endpoint (for both fips and non-fips).
Install and update scripts now use instance metadata to get the
endpoints.

 * Unit Tests : Y
 * Integration Tests : Tested sample deployment in us-west-2. The S3 and
   CodeDepoly endpoints were tested in a Government region. Still waiting on
   ISO region testing to make sure that everything works.