[Bug Fix] Remove logging statements in install and update scripts inside IMDSv2 class
Prior to this change, the logging statements inside the IMDSv2 class were being nil for some reason, which caused the install and update scripts to fall back to pulling the agent from the us-east-1 default region, which is not the expected behavior.

This change removes those log statements so we don't hit that error.

 * Unit Tests : [Y]
 * Integration Tests : Manually spun up an instance, caused the IMDSv2 call to fail by changing the endpoint for the token, and the install script worked as expected.