[Bug Fix] Upgrade aws-sdk-code-generator dependency to 0.2.2.pre
Prior to this change, there was a bug in the aws-sdk-code-generator ruby gem that caused it to not work on Windows. The fix for that issue was put up and then a new version of the gem was released that included that.

This change updates our dependency on aws-sdk-code-generator to use the now fixed aws-sdk-code-generator gem.

 * Unit Tests : [Y]
 * Integration Tests : [N/A]