IMDS service is no-longer assumed to be available
IMDS provides a wide range of host and data properties that can be used to setup and run the service. However, it is not always available due to security constraints or as the agent is running from an on-prem host. This updates allows for the situation where IMDS is not available and provides reasonable fallbacks.