handle unzip error when disk is full
http://infozip.sourceforge.net/FAQ.html#error-codes

During DownloadBundle, if unzip exit code is 50, that means disk
is/was full. If the child thread attempts to use Ruby's unzip to
extract again, the OS will signal to kill the child process.

When master process spawns a new child, the new child does not
pick up the DownloadBundle event, then the deployment is stuck
due to no response from agent.

This commit detects the unzip's exit code 50 of full disk, removes
partially extracted files, and raise an exception to post
failure to CodeDeploy server. The error message will be visible
as a lifecycle event error message, and the host-level deployment
will stop without being stuck/timedout.