Implement `is_noop?` for `HookExecutor`
This patch implements a simple noop check
for the `HookExecutor`. The logic is
essentially copied from `execute` a
few lines below, so the code
(and associated tests) are quite trivial.

I did have to make a decision around
handling scripts that do not exist -
this gets a bit into the weeds of how
we define a noop command. `execute`
will not run a script if it cannot be
found at the specified path, which
means that any command for which all
scripts do not exist are technically
noops. However, I think that customers
would rather not have us hit
`PutHostCommandComplete` in the rare
case that we receive a `Failed` for
a lifecycle event that has scripts
specified but none of them are valid
paths.