Add `is_command_noop?` for `CommandExecutor`
This patch adds a new function onto the
`CommandExecutor` which checks if all of
the command's lifecycle events are noops.
At runtime, the `CommandExecutor` uses
metaprogramming to set up methods for each
command that create a new `HookExecutor`
for each of the command's lifecycle events.

The agent appears to have support for
one-to-many mappings of commands to
lifecycle events, but we appear to only use
"identity mappings".  However, just to
be safe, I follow a similar pattern as the
`map` function does that handles this
one-to-many case.