Add retry with exponential backoff for install script while downloading version file from S3
Why is this change needed?
---------
Prior to this change, install script sometimes fails when S3 connection is timedout while
downloading version file. This happens when there is a new region build. But when the s3 download
is retried it works fine. so added retry logic to not fail in the first attempt.

How does it address the issue?
---------
This change added retry logic to not fail in the first attempt.