Uncomment windows specific unit test
Prior to this change the test in winagent_test.rb was commented out.

When we run bb on our dev desktops, the test task in hte Rakefile runs the winagent_test.rb and fails because the test runs on a linux machine
In this commit we uncomment the test and remove it from the list of files run as part of the "test" task.
I will add a new task in a followup commit to run windows specific tests